Move server to app
